+++ packages/terra-form-field/CHANGELOG.md
@@ -2,6 +2,11 @@
 
 ## Unreleased
 
+* Changed
+  * Upgrade tests to terra-functional-testing and WebDriverIO v6
+  * Removed unnecessary `wdio.conf.js` file
+  * Updated Jest command
+
 ## 4.15.0 - (February 2, 2021)
 
 * Changed
@@ -11,7 +16,7 @@
 
 * Fixed
  * Fixed broken links in documentation.
- 
+
 ## 4.13.0 - (December 8, 2020)
 
 * Changed